BROKEN everything is broken
okay hear me out on this: I know everything is broken, and it was sort of working, but I basically ported every function to use the new models/structs, which obviously broke a lot of stuff. But it does compile, so they're at least runtime errors that I can (and will) fix.